requirements.txt
colorlog==6.9.0
PyYAML==6.0.2
